Changed $scope_allowed to include "weekly"
Previously it was not sending to "weekly" people unless the current day is Monday. But that means if no-one visits the site on Monday it will not be sent out. Now, it will always try to send which is OK because it has the checks in order not to send if the person has received it since last Monday